refactor(front end): refactored common code between DisplayCards and PreviewDimensions
Refactored common code between DisplayCards and PreviewDimensions
