+++ src/types/card.ts
@@ -1,7 +1,12 @@
-export interface Card {
+export type Card = {
   // id: string;
   // dimensionID: string;
   statement: string;
   anchorSliderPos: number;
   isEditing: boolean;
+};
+
+export enum CardSide {
+  Left,
+  Right,
 }
